Removed the section on signing off commits.
Although the end idea of this process is an idea I fully support it doesn't add
much value unless I'm running some multi million Euro open source project. I'll
still include DCOs (e.g. in the case of ruby-lint) because as far as I'm aware
these are still valid even when not signing off commits (they make no mention
of this being required).

In the end not requiring this just makes the lifes of both me and contributors
easier as well as making it less daunting to contribute to my code.