Use a slightly lighter shade of black for text
This makes it a teeny bit easier one the eyes with the white
background. While I would like a gray background similar to my NeoVim
theme, I suspect this may look a bit off depending on the display
calibration used by visitors, so best to stick with just a white
background.