Generate video poster and rip out lazy loading
Lazy loading images is a nice idea on paper, but it gets messy as you
need to know the exact sizes of images and then have to make sure that
`max-width: 100%` doesn't mess up the aspect ratio. Fixing that in turn
seems to re-introduce some content reflowing, so let's just get rid of
lazy loading entirely.